+++ api/Controller/StockController.cs
@@ -19,14 +19,20 @@ namespace api.Controller
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            
             var stocks = await _stockRepo.GetAllAsync();
             var stockDto = stocks.Select(s => s.ToStockDto());
             return Ok(stockDto);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById(int id) 
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+            
             var stock = await _stockRepo.GetByIdAsync(id);
             if (stock == null)
             {
@@ -38,15 +44,21 @@ namespace api.Controller
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] StockCreateDto stockCreateDto)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+            
             var stockModel =  stockCreateDto.CreateStockDtoToStock();
             await _stockRepo.CreateAsync(stockModel);
             return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel.ToStockDto()); 
         }
 
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> UpdateAsync([FromRoute] int id, [FromBody] StockUpdateDto stockUpdateDto)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+            
             var stockModel = await _stockRepo.UpdateAsync(id, stockUpdateDto);
             if (stockModel == null)
             {
@@ -56,9 +68,12 @@ namespace api.Controller
         }
 
         [HttpPut]
-        [Route("{id}/{CompanyName}")]
+        [Route("{id}:int/{CompanyName}:string")]
         public async Task<IActionResult> UpdateCompanyAsync([FromRoute] int id, [FromRoute] StockUpdateDto stockUpdateDto)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+            
             var stockModel = await _stockRepo.UpdateCompanyAsync(id, stockUpdateDto);
             if (stockModel == null)
             {
@@ -68,9 +83,12 @@ namespace api.Controller
         }
 
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id}:int")]
         public async Task<IActionResult> DeleteAsync(int id)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+            
             var stockModel = await _stockRepo.DeleteAsync(id);
             if (stockModel == null)
             {
